Refactor XORTrialCustomEval into OneVarTrialCustomEval
